completed automated tests for business policy search

+++ src/main/scala/com/horizon/exchangeapi/ApiUtils.scala
@@ -164,7 +164,7 @@ object ApiTime {
   def nowUTC = ZonedDateTime.now.withZoneSameInstant(ZoneId.of("UTC")).toString
 
   /** Return UTC format of the time specified in seconds */
-  def thenUTC(seconds: Int) = ZonedDateTime.ofInstant(Instant.ofEpochSecond(seconds), ZoneId.of("UTC")).toString
+  def thenUTC(seconds: Long) = ZonedDateTime.ofInstant(Instant.ofEpochSecond(seconds), ZoneId.of("UTC")).toString
 
   /** Return UTC format of the time n seconds ago */
   def pastUTC(secondsAgo: Int) = ZonedDateTime.now.minusSeconds(secondsAgo).withZoneSameInstant(ZoneId.of("UTC")).toString
